requirements.txt
keyring~=25.6
pyperclip~=1.9
python-gitlab~=5.6
